% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caloric2020.R
\name{caloric2020}
\alias{caloric2020}
\title{Access the caloric 2019 single-cell RNA-seq data}
\usage{
caloric2020(
  tissues = "All",
  processedCounts = FALSE,
  reducedDims = TRUE,
  infoOnly = FALSE
)
}
\arguments{
\item{tissues}{A character vector with the tissues to retrieve objects for.
A list of available tissues can be obtained using
\code{listTabulaMurisSenisTissues("Droplet")}.}

\item{processedCounts}{Logical scalar. If \code{TRUE}, include the processed
counts in addition to the raw counts in the SingleCellExperiment object.}

\item{reducedDims}{Logical scalar. If \code{TRUE}, include the PCA, tSNE
and UMAP representations in the SingleCellExperiment object (the tSNE
representation is not available for the full dataset ('All' tissue)).}

\item{infoOnly}{Logical scalar. If \code{TRUE}, only print the total size
of the files that will be downloaded to and/or retrieved from the cache.}
}
\value{
If \code{infoOnly} is \code{FALSE}, returns a named list of
\linkS4class{SingleCellExperiment} objects (one per tissue requested
via \code{tissues}). Otherwise, each element in the list is \code{NULL}.
}
\description{
Access the raw (10x Genomics) RNA-seq data from the caloric restriction study
by Ma et al. 2020
}
\details{
The data set was downloaded from GEO assession number GSE137869
(https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE137869).
}
\examples{
if (interactive()) {
  sce <- TabulaMurisSenisDroplet(tissues = "All")
}

}
\references{
Schaum et al (2019): The murine transcriptome reveals global aging nodes with
organ-specific phase and amplitude. bioRxiv doi:10.1101/662254.

The Tabula Muris Consortium (2020): A single-cell transcriptomic atlas
characterizes ageing tissues in the mouse. Nature 583:590–595.
}
\author{
Charlotte Soneson
}
