% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWebData.R
\name{getWebData}
\alias{getWebData}
\title{Get data from web repository via a URL link}
\usage{
getWebData(format, url, destination, gzipped = TRUE, sep = "\\t", header = T)
}
\arguments{
\item{format}{A character string. Describes input file format, "tsv", "csv", or "mtx".}

\item{url}{A character string. Full URL to file hosted on a web repository.}

\item{destination}{A character string. Local destination to download file.}

\item{gzipped}{Boolean. Describes whether web file is gzipped. Defaults to TRUE.}

\item{sep}{The field separator character. Follow values documented in read.table function.}

\item{header}{A logical value indicating whether the file contains the names of the variables as its first line. Again, refer to read.table function.}
}
\value{
A data frame or matrix.
}
\description{
Get data from web repository via a URL link
}
\examples{
test <- getWebData(format = "tsv", destination = "~/Desktop/",
url = "https://ftp.ncbi.nlm.nih.gov/geo/series/GSE222nnn/GSE222510/suppl/GSE222510_processed_barcodes.tsv.gz",
header=F)
head(test)
}
