% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parabiosis_barcode.R
\docType{data}
\name{parabiosis_barcode}
\alias{parabiosis_barcode}
\title{Barcodes for Parabiosis Filtered Dataset}
\format{
\subsection{A data frame with 1 column and 105329 rows:}{

\describe{
\item{barcode}{unique cell barcode.}
}
}
}
\source{
\url{https://www.nature.com/articles/s43587-023-00373-6}
}
\usage{
parabiosis_barcode
}
\description{
Unique cell barcodes for parabiosis filtered 10X Genomics dataset.
}
\examples{
data(parabiosis_barcode) # Lazy loading
}
\keyword{datasets}
