% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parabiosis_genes.R
\docType{data}
\name{parabiosis_genes}
\alias{parabiosis_genes}
\title{Genes for Parabiosis Filtered Dataset}
\format{
\subsection{A data frame with 1 column and 20905 rows:}{

\describe{
\item{genes}{parabiosis genes}
}
}
}
\source{
\url{https://www.nature.com/articles/s43587-023-00373-6}
}
\usage{
parabiosis_genes
}
\description{
20905 Genes in ENSEMBL format for the filtered 10X Genomics Parabiosis Mouse Study
}
\examples{
data(parabiosis_genes) # Lazy loading
}
\keyword{datasets}
