% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parabiosis_metadata}
\alias{parabiosis_metadata}
\title{Parabiosis Data}
\format{
\subsection{parabiosis_metadata}{

A data frame with 105329 rows and 12 columns:
\describe{
\item{barcode}{cell barcode}
\item{nCount_RNA}{number of RNA transcripts}
\item{animal name}{isochronic or parabionic conditions}
...
}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE222510}
}
\usage{
parabiosis_metadata
}
\description{
Filtered metadata for 10X Genomics parabiosis experiment
}
\examples{
data(parabiosis_metadata) # Lazy loading
}
\keyword{datasets}
