% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parabiosis_metadata.R
\docType{data}
\name{parabiosis_metadata}
\alias{parabiosis_metadata}
\title{Metadata for Filtered Dataset of Parabiosis Mouse Aging Project}
\format{
\subsection{A data frame with 12 columna and 105329 rows:}{

\describe{
\item{\code{cell}:}{Character, unique cell identifier across all samples.}
\item{\code{orig.ident}:}{Character, origin project identifier.}
\item{\code{nCount_RNA}:}{Numeric, number of RNA transcripts.}
\item{\code{nFeature_RNA}:}{Integer, number of RNA features.}
\item{\code{sample_order}:}{Character, unique animal identifier and sequencing batch.}
\item{\code{animal}:}{Character, unique animal identifier, includes heterochronic or isochronic condition.}
\item{\code{batch}:}{Character, sequencing batch identifier.}
\item{\code{animal_name}:}{Character, isochronic or heterochronic conditions.}
\item{\code{percent.mito}:}{Numeric, percentage of mitochondrial content.}
\item{\code{percent.ribo}:}{Numeric, percentage of ribosomal content.}
\item{\code{cell_type}:}{Character, cell type to which the cell was assigned.}
\item{\code{subpopulation}:}{Character, cell subpopulation to which cell was assigned.}
}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE222510}
}
\usage{
parabiosis_metadata
}
\description{
Filtered metadata for 10X Genomics parabiosis experiment
20905 Genes in ENSEMBL format and 105329 cells
}
\examples{
data(parabiosis_metadata) # Lazy loading
}
\keyword{datasets}
